完善了 90% 的全局 temp Signed-off-by: Gavin <meiyan532484710@163.com>

+++ core/ppos_storage/ppos_storageCache.go
@@ -43,7 +43,7 @@ type ticketDependency struct {
 	// ticket age
 	Age *big.Int
 	// ticket count
-	Num uint64
+	Num uint32
 	// ticketIds
 	Tids []common.Hash
 }
@@ -70,7 +70,7 @@ func (td *ticketDependency) SubNum() {
 
 type ticket_temp struct {
 	// total remian  k-v
-	Sq int
+	Sq int32
 	// ticketInfo  map[txHash]ticketInfo
 	Infos map[common.Hash]*types.Ticket
 	// ExpireTicket  map[blockNumber]txHash
